git-svn-id: https://svn.esynctraining.com/svn/esynctraining@14884 41667c05-fd31-8a45-9ea7-b4ae18c5af78

+++ EdugameCloud.Lti/API/AdobeConnect/MeetingSessionService.cs
@@ -56,7 +56,7 @@ namespace EdugameCloud.Lti.API.AdobeConnect
             var listOfEvents = new List<MeetingSessionDTO>();
             while (startDate < latestDateToCheck)
             {
-                if (dto.DaysOfWeek.Contains(startDate.DayOfWeek))
+                if (dto.DaysOfWeek.Contains((int)startDate.DayOfWeek))
                 {
                     var ev = new MeetingSessionDTO
                     {
